Time: 0 ms (100%), Space: 7.6 MB (91.43%) - LeetHub

+++ 0693-binary-number-with-alternating-bits/0693-binary-number-with-alternating-bits.cpp
@@ -15,7 +15,7 @@ public:
         // return true;
 
         //Solution 2: Usin xor bit manipulation
-        unsigned int x = n ^ (n >> 1);
+        long x = n ^ (n >> 1);
         return (x & (x+1)) == 0;
     }
 };
